+++ main/src/ca/uwaterloo/flix/language/phase/jvm/JvmBackend.scala
@@ -78,7 +78,7 @@ object JvmBackend {
 
     val unitClass = List(JvmClass(BackendObjType.Unit.jvmName, BackendObjType.Unit.genByteCode()))
 
-    val flixErrorClass = List(JvmClass(BackendObjType.FlixError.jvmName, BackendObjType.FlixError.genByteCode()))
+    val flixErrorClass = List(JvmClass(JvmName.FlixError, ClassMaker.FlixError.genByteCode()))
     val rslClass = List(JvmClass(BackendObjType.ReifiedSourceLocation.jvmName, BackendObjType.ReifiedSourceLocation.genByteCode()))
     val holeErrorClass = List(JvmClass(BackendObjType.HoleError.jvmName, BackendObjType.HoleError.genByteCode()))
     val matchErrorClass = List(JvmClass(BackendObjType.MatchError.jvmName, BackendObjType.MatchError.genByteCode()))
